https://forum.pcmtec.com/topic/1491-micro-controller-using-can-bus-data-for-ancillary-control/
I completed a manual conversion on my BA Mk2 XR6T, for some reason, the PCM would not ground the reverse lockout wire. I understand that the PCM is supposed to ground the pin when the car is travelling less than 7km/h and the clutch is depressed. I read that most people use the brake light circuit to ground the wire but I didn't like this idea.

The data I required is available via the CAN bus so I bought a micro-controller with a native CAN bus interface and put together a prototype board that does the following:

Listens on the CAN bus for the vehicle speed.
When the vehicle speed is below 5km/h, once per second, sends a query to the PCM asking the status of the clutch pedal.
When the vehicle speed and clutch status are correct, pulls the reverse lockout wire to ground using a MOSFET (and turn on an LED to show me it's working).